boot/recovery-menu: extract reboot modes
