+++ boot/recovery-menu/main.rb
@@ -7,23 +7,6 @@ def run(*cmd)
   system(*cmd) unless LVGL::Introspection.simulator?
 end
 
-module RebootModes
-  Android = {
-    "recovery"   => ["Reboot to recovery", ->() { run("reboot recovery") }],
-    "bootloader" => ["Reboot to bootloader", ->() { run("reboot bootloader") }],
-  }
-
-  def self.options()
-    [
-      ["Reboot to system", ->() { run("reboot") }],
-    ] + 
-      Configuration["HAL"]["boot"]["rebootModes"].map do |identifier|
-        const, key = identifier.split(".", 2)
-        const_get(const)[key]
-      end
-  end
-end
-
 module BootGUI
   class MainWindow < LVGUI::BaseWindow
     include LVGUI::ButtonPalette
